backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Anthropic API
anthropic>=0.40.0

# Utilities
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# WebSocket
websockets==12.0

# Data generation
faker==22.5.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
